LayoutTests/fast/dom/qualified-name-validation-expected.txt

Tests failing cases for qualified name validation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.createElement('@foo') threw exception InvalidCharacterError: The string contains invalid characters..
PASS document.createElementNS('ns', '@foo') threw exception InvalidCharacterError: Invalid qualified name start in '@foo'.
PASS document.createElementNS('ns', 'f@ast') threw exception InvalidCharacterError: Invalid qualified name part in 'f@ast'.
PASS document.createAttribute('@foo') threw exception InvalidCharacterError: Invalid qualified name: '@foo'.
PASS document.createAttributeNS('ns', '@foo') threw exception InvalidCharacterError: Invalid qualified name start in '@foo'.
PASS document.createAttributeNS('ns', 'f@ast') threw exception InvalidCharacterError: Invalid qualified name part in 'f@ast'.
PASS document.body.setAttribute('@foo', 'test') threw exception InvalidCharacterError: Invalid qualified name: '@foo'.
PASS document.body.setAttributeNS('ns', '@foo', 'test') threw exception InvalidCharacterError: Invalid qualified name start in '@foo'.
PASS document.body.setAttributeNS('ns', 'f@st', 'test') threw exception InvalidCharacterError: Invalid qualified name part in 'f@st'.
PASS document.createProcessingInstruction('@foo', 'test') threw exception InvalidCharacterError: Invalid qualified name: '@foo'.
PASS successfullyParsed is true

TEST COMPLETE
